hidock-desktop-app/requirements.txt
pyusb
customtkinter
Pillow
pygame
google-generativeai
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<1.12.0
pydub
matplotlib>=3.5.0,<3.8.0
librosa>=0.8.0
pyparsing>=3.0.0,<4.0.0

# Development and testing dependencies
pytest
pytest-cov
pytest-mock
pytest-asyncio
black
flake8
isort
pylint
mypy
